Att historic model and controller created and type in actives SELL and BUY implemented

+++ controller/ativoController.js
@@ -1,39 +1,81 @@
 const ativoModel = require('../model/ativoModel');
+const historicoController = require('./historicoController');
 
-const listarAtivos = (req, res) => {
+const listarAtivos = async (req, res) => {
 
-    ativoModel.findAll({raw: true, order:[['id','DESC']]}).then((ativos) => {
-        res.json(ativos)
+    await ativoModel.findAll({raw: true, order:[['id','DESC']]}).then((ativos) => {
+        if(ativos !== null)
+            res.json(ativos);
+        else
+            res.status(200).send('Não existem ativos.');
+    }).catch(() => {
+        res.status(500).send('Internal server error');
     })
 }
 
-const criarAtivo = (req, res) => {
+const criarAtivo = async (req, res) => {
  
-    ativoModel.create({
+    if(Object.keys(req.body).length === 0)
+    {
+        res.status(400).send('Corpo da requisição vazio');
+        return;
+    }
+
+    await ativoModel.create({
         tipoativo: req.body.tipoativo,
         codigoativo: req.body.codigoativo,
         datacompra: req.body.datacompra,
         quantidade: req.body.quantidade,
         preco: req.body.preco,
-        outroscustos: req.body.outroscustos
+        outroscustos: req.body.outroscustos,
+        tipo: req.body.tipo,
+        usuarioID: req.body.usuarioID
     }).then((data) => {
-        res.json("Ativo Criado");
-    })
-}
-
-const excluirAtivo = (req, res) => {
-
-    ativoModel.destroy({where:{id: req.params.id}}).then(() => {
-        res.json("Ativo Excluido")
         
+        historicoController.criarHistorico(req, res)
+
+        res.status(200).send("Ativo Criado");
+    }).catch(() => {
+        res.status(500).send('Internal server error');
     })
+
+    
 }
 
-const alterarAtivo = (req, res) => {
+const excluirAtivo = async (req, res) => {
+
+    // Verifica se o id passado existe
+    const verifyID = await ativoModel.count({where:{id:req.params.id}})
+
+    if(verifyID)
+    {
+        ativoModel.destroy({where:{id: req.params.id}}).then(() => {
+            res.json("Ativo Excluido")
+        }).catch(() => {
+            res.status(500).send("internal server error")
+        })
+    }
+    else
+    {
+        res.status(200).send("Ativo não encontrado");
+    }
+}
+
+const alterarAtivo = async (req, res) => {
+
+    // Verifica se o id passado existe
+    const verifyID = await ativoModel.count({where:{id:req.params.id}})
     
-    ativoModel.update(req.body, {where:{id: req.params.id}}).then(() => {
-        res.json("Ativo Alterado")
-    })
+    if(verifyID)
+    {
+        ativoModel.update(req.body, {where:{id: req.params.id}}).then(() => {
+            res.status(200).send('Ativo foi alterado.')
+        })
+    }
+    else
+    {
+        res.status(200).send("Ativo não encontrado");
+    } 
 }
 
 module.exports = {listarAtivos, criarAtivo, excluirAtivo, alterarAtivo}
